requirements.txt
torch==2.0.1
torchvision==0.15.2
albumentations==1.3.1
opencv-python==4.9.0.80
Pillow==10.3.0
Flask==2.2.5
numpy==1.24.4
